Cleaned up for cross compilation.

+++ serial_posix.go
@@ -173,11 +173,9 @@ func newTermios(c *Config) (termios *syscall.Termios, err error) {
 	}
 	termios.Cflag |= flag
 	// Input baud.
-	//termios.Ispeed = flag
-	cfSetIspeed(termios, uint64(flag))
+	cfSetIspeed(termios, flag)
 	// Output baud.
-	//termios.Ospeed = flag
-	cfSetOspeed(termios, uint64(flag))
+	cfSetOspeed(termios, flag)
 	// Character size.
 	if c.DataBits == 0 {
 		flag = syscall.CS8
